requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
reportlab==4.0.0
fpdf==1.7.2
